requirements.txt
aiofiles==22.1.0
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.19.0
ansiwrap==0.8.4
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.2
async-timeout==4.0.2
attrs==22.2.0
Babel==2.11.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.11.1
black==23.3.0
bleach==6.0.0
boto3==1.26.133
botocore==1.29.133
Brotli==1.0.9
cattrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
comm==0.1.2
contextvars==2.4
contourpy==1.0.7
coverage==7.1.0
cursor==1.3.5
cycler==0.11.0
dao-analyzer==1.2.6
dash==2.5.1
dash-bootstrap-components==1.3.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataframe-image==0.1.10
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
docstring-to-markdown==0.11
entrypoints==0.4
exceptiongroup==1.1.0
executing==1.2.0
fastjsonschema==2.16.2
flake8==6.0.0
Flask==2.1.3
Flask-Compress==1.13
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.30
gql==3.5.0b0
graphql-core==3.3.0a2
halo==0.0.29
idna==3.4
immutables==0.19
IneqPy @ git+https://github.com/Grasia/ineqpy@570512dad0d6c3deaf102c306939c9b55a90b1ee
iniconfig==2.0.0
ipyflow==0.0.189
ipyflow-core==0.0.189
ipykernel==6.23.1
ipython==8.8.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
isort==5.11.4
itsdangerous==2.1.2
jedi==0.18.2
jedi-language-server==0.40.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter-server-mathjax==0.2.6
jupyter-ydoc==0.2.4
jupyter_client==8.2.0
jupyter_core==5.1.5
jupyter_server==2.6.0
jupyter_server_fileid==0.9.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.8.0
jupyterlab==4.0.5
jupyterlab-execute-time==3.0.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
jupyterlab_git==0.42.0
jupyterlab_server==2.19.0
kaleido==0.2.1
kiwisolver==1.4.4
llvmlite==0.40.0
log-symbols==0.0.14
lsprotocol==2022.0.0a10
markdownify==0.11.6
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
mccabe==0.7.0
millify==0.1.1
mistune==2.0.4
multidict==6.0.4
mypy==0.991
mypy-extensions==0.4.3
nb-search==1.0.8
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.9
nbdime==3.1.1
nbformat==5.7.3
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numba==0.57.0
numpy==1.23.5
overrides==7.3.1
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
papermill==2.4.0
parso==0.8.3
pathspec==0.11.1
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.6.2
plotly==5.13.0
plotly-calplot==0.1.13
pluggy==1.0.0
portalocker==2.7.0
prometheus-client==0.16.0
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyarrow==10.0.1
pyccolo==0.0.49
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.4
pyflakes==3.0.1
pygls==1.0.0
Pygments==2.14.0
pyparsing==3.0.9
pyrsistent==0.19.3
pytest==7.2.1
pytest-cov==4.0.0
python-dateutil==2.8.2
python-json-logger==2.0.4
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.3
QtPy==2.3.1
requests==2.28.2
requests-cache==0.9.8
requests-toolbelt==0.10.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rolling==0.5.0
s3transfer==0.6.1
scikit-learn==1.2.1
scipy==1.10.0
seaborn==0.12.2
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
spinners==0.0.24
stack-data==0.6.2
statsmodels==0.13.5
tabulate==0.8.10
tenacity==8.1.0
termcolor==2.3.0
terminado==0.17.1
textwrap3==0.9.2
threadpoolctl==3.1.0
thrift==0.13.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tornado==6.3.2
tqdm==4.64.1
traitlets==5.8.1
typeguard==2.13.3
typing_extensions==4.4.0
tzdata==2023.3
uri-template==1.2.0
url-normalize==1.4.3
urllib3==1.26.14
vulture==2.7
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.0.3
widgetsnbextension==4.0.5
y-py==0.5.9
yarl==1.8.2
ypy-websocket==0.8.2
